[Storage] Move TrimPolicy to api.py; restore l2_orig_indices naming
Relocate the TrimPolicy enum from prefetch_controller into the leaf
api module so trace codecs (and other type consumers) can import it
without pulling the prefetch_controller dependency chain — breaking a
latent codecs<->prefetch_controller import cycle. Revert the
PrefetchHandle.l2_orig_indices field name (was l2_found_indices).
